Major Update, Fetch online files
Changed from using set /p commands to CHOICE so that the prompts are less prone to breaking if an invalid character is pressed. Compartmentalized the Chrome, Adobe, and Java installers. Changed the install processes to fetch online files if they are not already downloaded. Changed the Java to use Microsoft JDK instead of Oracle. Updated the Windows Update process to be a little bit smoother.